Change Simualtion model IntegerFields to FloatFields Add logic for changing POST request in simulation form depending on checkbox state

+++ mastersheet/models.py
@@ -18,19 +18,19 @@ class Simulation(models.Model):
     undertray_name = models.CharField(max_length=10)
     nose_name = models.CharField(max_length=10)
 
-    front_wing_df = models.IntegerField()
-    rear_wing_df = models.IntegerField()
-    sidepod_df = models.IntegerField()
-    diffuser_df = models.IntegerField()
-    undertray_df = models.IntegerField()
-    nose_df = models.IntegerField()
+    front_wing_df = models.FloatField()
+    rear_wing_df = models.FloatField()
+    sidepod_df = models.FloatField()
+    diffuser_df = models.FloatField()
+    undertray_df = models.FloatField()
+    nose_df = models.FloatField()
 
-    front_wing_drag = models.IntegerField()
-    rear_wing_drag = models.IntegerField()
-    sidepod_drag = models.IntegerField()
-    diffuser_drag = models.IntegerField()
-    undertray_drag = models.IntegerField()
-    nose_drag = models.IntegerField()
+    front_wing_drag = models.FloatField()
+    rear_wing_drag = models.FloatField()
+    sidepod_drag = models.FloatField()
+    diffuser_drag = models.FloatField()
+    undertray_drag = models.FloatField()
+    nose_drag = models.FloatField()
 
     def __str__(self):
         return self.chassis_name
